[NFC][Object] Move BBAddrMap related types to a shared header
This is part of patches to port BBAddrMap to COFF.

Introduce BBAddrMap.h and move BBAddrMap/PGOAnalysisMap type definitions
out of ELFTypes.h.
